# api/openapi.yaml
openapi: 3.0.3
info:
  title: GoSheet API
  description: |
    HTTP API for GoSheet spreadsheet application.

    ## Authentication

    All `/api/*` endpoints require a bearer token. Two token types are in use:

    **Bootstrap token** (`BootstrapAuth`): Generated by the Go server at startup using a 32-byte
    CSPRNG (base64url encoded). Written to fd 3 alongside the port number. Electron injects it into
    the renderer, which sends it as `Authorization: Bearer <token>` on every request. Has full
    access to all endpoints. In `NODE_ENV=test` mode the server skips token generation and auth is
    disabled entirely so Playwright tests run without modification.

    **Agent token** (`AgentAuth`): Scoped token (`ro` or `rw`) issued by `POST /api/agent/token`
    in exchange for a valid bootstrap token. Used exclusively by AI agents. Agent tokens are
    rejected on `/api/file/*` endpoints regardless of scope. Read-only (`ro`) tokens are rejected
    on any mutating agent endpoint.

    The bootstrap endpoint (`GET /api/agent/bootstrap`) also accepts the agent token as a
    `?token=` query parameter for copy-paste UX when the user hands a URL to an agent.
  version: 1.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: Relative to host (localhost ephemeral port)

# All /api/* endpoints require auth. Static file routes are unauthenticated.
# In NODE_ENV=test the server disables auth entirely (bootstrapToken == "").
security:
  - BootstrapAuth: []
components:
  securitySchemes:
    BootstrapAuth:
      type: http
      scheme: bearer
      description: |
        Bootstrap token generated at server startup. Injected into the Electron renderer via URL
        query param and forwarded as `Authorization: Bearer <token>` on every API request.
        Has full access to all endpoints including `/api/file/*`.
    AgentAuth:
      type: http
      scheme: bearer
      description: |
        Scoped agent token issued by `POST /api/agent/token`. Used by AI agents.
        Scope `rw` allows reads and writes; scope `ro` allows reads only.
        Rejected on `/api/file/*` endpoints regardless of scope.
  schemas:
    ErrorResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code (e.g. INVALID_CELL_REF, FILE_NOT_FOUND)
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
        code:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    PathRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
    SetCellRequest:
      type: object
      required: [row, col, value]
      properties:
        row:
          type: integer
        col:
          type: integer
        value:
          type: string
    CellValueResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                computed:
                  type: string
...
    SetCellResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                value:
                  type: string
                displayValue:
                  type: string
                isFormula:
                  type: boolean
                hasUnsavedChanges:
                  type: boolean
...
    SaveLoadResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            path:
              type: string
...
    MergeRegion:
      type: object
      required: [startRow, startCol, rowSpan, colSpan]
      properties:
        startRow:
          type: integer
        startCol:
          type: integer
        rowSpan:
          type: integer
        colSpan:
          type: integer
...
    MergesResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                merges:
                  type: array
                  items:
                    $ref: '#/components/schemas/MergeRegion'
...
paths:
  /api/cell/value:
    get:
      operationId: getCellValue
      summary: Get cell computed value
      parameters:
        - name: row
          in: query
          required: true
          schema:
            type: integer
        - name: col
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cell value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CellValueResponse'
        '400':
          description: Bad request (invalid or missing row/col)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/cell/set:
    post:
      operationId: setCellValue
      summary: Set cell value
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCellRequest'
      responses:
        '200':
          description: Cell updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetCellResponse'
        '400':
          description: Bad request (invalid body, missing fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent token with ro scope, or agent token on file endpoint)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error (e.g. circular ref, invalid formula)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cell/ref:
    get:
      operationId: getCellRef
      summary: Get cell reference (e.g. A1)
      parameters:
        - name: row
          in: query
          required: true
          schema:
            type: integer
        - name: col
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cell reference string
          content:
            application/json:
              schema:
                type: string
                example: '"A1"'
        '400':
          description: Bad request (invalid or missing row/col)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/merges:
    get:
      operationId: getMerges
      summary: Get all merge regions
      responses:
        '200':
          description: Merge regions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MergesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/file/save:
    post:
      operationId: saveFile
      summary: Save spreadsheet to file
      description: Requires bootstrap token. Agent tokens are rejected on all `/api/file/*` endpoints.
      security:
        - BootstrapAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
      responses:
        '200':
          description: File saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveLoadResponse'
        '400':
          description: Bad request (missing path)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent tokens cannot perform file operations)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error (e.g. file write error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/file/load:
    post:
      operationId: loadFile
      summary: Load spreadsheet from file
      description: Requires bootstrap token. Agent tokens are rejected on all `/api/file/*` endpoints.
      security:
        - BootstrapAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathRequest'
      responses:
        '200':
          description: File loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveLoadResponse'
        '400':
          description: Bad request (missing path)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent tokens cannot perform file operations)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error (e.g. file not found, read error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/file/new:
    post:
      operationId: newFile
      summary: Create new empty spreadsheet
      description: Requires bootstrap token. Agent tokens are rejected on all `/api/file/*` endpoints.
      security:
        - BootstrapAuth: []
      responses:
        '200':
          description: New file created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (agent tokens cannot perform file operations)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...